Update to support Spring 5.x and above.
Motivation
----------
The current library supports a baseline of Spring 3.x which is quite old and no longer supported. It also has a JDK baseline of JDK 1.4 which is far out of date. Since Spring 5.x the JdbcTemplate implementation has removed the NativeJdbcExtractor interface as it is no longer needed with JDBC 4.0 drivers and the unwrap(Class<?>) method. Because of this breaking change, this library no longer functions with Spring 5.x and above.

Modifications
-------------
The dependencies defined in the POM file have been brought up to their latest stable versions. This includes moving Spring to 5.1.x and the test dependencies to the newest versions. The JDK target version has also been updated to JDK 8 as it is the oldest supported version supported by Spring. Finally, the code has been updated to remove usages of the NativeJdbcExtractor as it is no longer needed.

Result
------
The library will now function on modern versions of Spring as well as remaining backward compatible with older versions (albeit without NativeJdbcExtractor support).